+++ src/error.rs
@@ -70,3 +70,14 @@ impl From<Utf8Error> for Error {
         Error::Utf8(err)
     }
 }
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[test]
+    fn error_send_and_sync() {
+        fn assert_send_sync<T: Send + Sync>() {}
+        assert_send_sync::<Error>();
+    }
+}